Force binding all dynamic bound symbols on macos.
Due to the cyclic dynamic library dependency,
the logical time symbols are resolved dynamically
during runtime. This appears to be racy on the
test machine I have access to.
This commit works around this race condition by
forcing all symbols to be bound at load time
of the rti library.