Do not normailze root object if they do not look like root objects.
This change tries to guess better if the object model given root
object is meant to be the root object or not. If you provide
a standard conformant object model file, this change aims to be
a noop. But there are some object models out there which seem to be
accepted by other implementations that require some kind of
'do what I think' behaviour so that we cannot just bail out
on non standard conformant object models.
So instead of just normalizing the first class to be the root class,
we try to insert an artificial root class if there is no such class.