Make librti1516e relocatible.
This aims to find the management object model relative to the
rti1516e library file. Thus you should be able to move the whole
installed file tree to wherever you want and it should still work.
If the operating system does not support querying the location
of the shared library, the traditional CMAKE_INSTALL_PREFIX
based approach is used.