Fix races with dynamic loading on macos.
It looks like we still need a solution for
the lazy reaolving of symbols on macos.
The previous approach implemented at
272f405967585985a56bc1efa1a236618a24b6f9
was later reverted with
ee7d3f662e7bfbf15d9f7834314f040290aedbd7.
This is now an attempt to use a non deprecated
solution for this problem.

Thanks to James Turner providing this fix!