Use a designated InsertModulesMessage to insert these.
The standard document states that the fom modules provided
on join persist and are available to all joined federates.
This is actually much easier to get right than what I
understood of object modules on join before. This change
should already provide all the ground work to implement
fom modules on join.