Try to work around systems where their linker default to --as-needed.
Still just a workaround. The problem arises again from the standardized
circular dependency of fedtime -> rti -> fedtime -> ... .
Cmake does not like these dependencies either and even better reorders
the libraries into the dependency order it knows about. But that one is
exactly the wrong one for our purpose. Even though even the currently
working link order may no longer work somewhere in the future if you
think about linking.
Now with system linkers that start to default to as-needed linking
the requirement to link in the fedtime library somewhere into the
final executable gets dropped as it is not directly needed.
So for now just disable the as needed stuff for our build.
Note that the package config files contain a link order that
should work as of today on as-needed systems.