Remove special arithmetic handling for small logical times.
Before we had some special implementation for the floating
point valued logical times to advance to a non equal next
logical time even if roundoff would result in the same
floating point value.
This change removes this special handling as this behavior
is probably unexpected for most of us.
The patch is kept completely separate so that we can easily
revert that change.