Fix exception specifications for newer c++ standards.
I did allways hope that we will observe an update to
the official standard how to treat deprecation of exception
specifications in c++11 and their removal in c++17. Since this
was not happening, fix this issue at least in OpenRTI.

With this change exception specifications will be removed
once c++17 or newer is selected in the compiler. Empty exception
specifications are replaced by noexcept for c++17 and newer.
Non empty exception specifications are translated into the empty
source code string.