Fix use of std::auto_ptr for newer c++ standards.
I did allways hope that we will observe an update to
the official standard how to treat deprecation of std::auto_ptr
in c++11 and its removal in c++17. Since this was not happening,
fix this issue at least in OpenRTI.

With this change std::auto_ptr is replaced with std::unique_ptr
once c++17 or newer is selected in the compiler.